requirements.txt
certifi==2018.8.13
chardet==3.0.4
Django==1.11.15
django-after-response==0.2.2
django-redis==4.9.0
djangorestframework==3.8.2
idna==2.7
PyAudio==0.2.11
pydub==0.22.1
pytz==2018.5
redis==2.10.6
requests==2.19.1
urllib3==1.23
